fix estilos, fix mensajes delete

+++ src/js/productos.js
@@ -146,12 +146,10 @@ document.querySelectorAll(".card").forEach((e) => {
   const eliminar = e.querySelector(".delete");
   eliminar.addEventListener("click", async () => {
     const idProd = e.dataset.id;
-    console.log("id en evento: ", idProd);
     const indLista = listaAPI.findIndex((producto) => producto.id === idProd);
-    console.log("id en lista, si existe: ", indLista);
     if (indLista !== -1) {
-      let respuesta = await conectionAPI.deleteItem(idProd);
-      if (respuesta.ok) {
+      if (confirm("¿Está seguro que desea eliminar?")) {
+        await conectionAPI.deleteItem(idProd);
         listaAPI.splice(idProd, 1);
         e.remove();
       }
